@@ -1,8 +1,9 @@
 package com.xuie.androidrecyclerview;
 
 import android.os.Bundle;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.xuie.androidrecyclerview.adapter.MultipleItemQuickAdapter;
 import com.xuie.androidrecyclerview.base.BaseActivity;
