@@ -1,8 +1,10 @@
 package com.xuie.recyclerview.callback;
 
 import android.graphics.Canvas;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
+
+import androidx.recyclerview.widget.ItemTouchHelper;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.View;
 
 import com.xuie.recyclerview.BaseItemDraggableAdapter;
